refactor(maps): extract useTabletopMapSync from TabletopView
Move the inbound tabletop-map party message reducer (apply token/text/drawing/
active-map changes to the query cache) into a useTabletopMapSync hook colocated
with the map cache helpers it uses. TabletopView now just calls the hook for its
`send` broadcaster. No behavior change; verified by the TabletopView unit test
and the active-map + drawing e2e specs.
